# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin API 文档",
        "title": "Gin Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "25076778@qq.com"
        },
        "version": "2.0"
    },
    "host": "127.0.0.1:8080",
    "paths": {
...
        "/api/v1/config-category": {
            "get": {
                "description": "配置分类列表",
                "tags": [
                    "配置分类管理"
                ],
                "summary": "列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "是否不分页",
                        "name": "notPage",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/request.PageData"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/model.ConfigCategory"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "配置分类创建",
                "tags": [
                    "配置分类管理"
                ],
                "summary": "创建",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "创建参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ConfigCategoryCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ConfigCategory"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/config-category/{id}": {
            "get": {
                "description": "配置分类详情",
                "tags": [
                    "配置分类管理"
                ],
                "summary": "详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ConfigCategory"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "配置分类更新",
                "tags": [
                    "配置分类管理"
                ],
                "summary": "更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ConfigCategoryUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/errcode.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "配置分类删除",
                "tags": [
                    "配置分类管理"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/errcode.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.ArgsErrorResponse"
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/cards": {
            "get": {
                "description": "仪表盘卡片统计 用户数/角色数/菜单数/文章数/今日日志/字典项/系统配置/导入记录",
                "tags": [
                    "仪表盘管理"
                ],
                "summary": "卡片统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/service.DashboardCard"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/statistics": {
            "get": {
                "description": "操作日志统计 方法统计/耗时分布/小时请求量/状态码/PV/UV/平均响应/峰值时段",
                "tags": [
                    "仪表盘管理"
                ],
                "summary": "操作日志统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.OperatorLogStatistics"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/dashboard/system-resource": {
            "get": {
                "description": "系统资源监控 CPU/内存/磁盘/带宽",
                "tags": [
                    "仪表盘管理"
                ],
                "summary": "系统资源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "认证Token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errcode.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.SystemResource"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "系统错误",
                        "schema": {
                            "$ref": "#/definitions/errcode.SystemErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errcode.ArgsErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Invalid arguments"
                }
            }
        },
        "errcode.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "errcode.SystemErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {},
                "msg": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
...
        "model.ConfigCategory": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "request.ConfigCategoryCreate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "request.ConfigCategoryUpdate": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "request.PageData": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "数据列表"
                },
                "notPage": {
                    "description": "不分页",
                    "type": "boolean",
                    "example": false
                },
                "page": {
                    "description": "当前页",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
...
        "service.CPUInfo": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "number"
                },
                "total": {
                    "type": "number"
                },
                "usage": {
                    "type": "number"
                }
            }
        },
        "service.CostDist": {
            "type": "object",
            "properties": {
                "bt100200": {
                    "type": "integer"
                },
                "bt200500": {
                    "type": "integer"
                },
                "bt50100": {
                    "type": "integer"
                },
                "gt500": {
                    "type": "integer"
                },
                "lt50": {
                    "type": "integer"
                }
            }
        },
        "service.DashboardCard": {
            "type": "object",
            "properties": {
                "change": {
                    "type": "integer"
                },
                "changePercent": {
                    "type": "number"
                },
                "count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "trend": {
                    "type": "string"
                },
                "yesterdayCount": {
                    "type": "integer"
                }
            }
        },
        "service.DashboardSummary": {
            "type": "object",
            "properties": {
                "avgCost": {
                    "type": "number"
                },
                "costChange": {
                    "type": "number"
                },
                "peakHour": {
                    "type": "integer"
                },
                "peakQps": {
                    "type": "number"
                },
                "pvChange": {
                    "type": "number"
                },
                "todayPv": {
                    "type": "integer"
                },
                "todayUv": {
                    "type": "integer"
                },
                "uvChange": {
                    "type": "number"
                },
                "yesterdayAvgCost": {
                    "type": "number"
                },
                "yesterdayPv": {
                    "type": "integer"
                },
                "yesterdayUv": {
                    "type": "integer"
                }
            }
        },
        "service.DiskInfo": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "string"
                },
                "total": {
                    "type": "string"
                },
                "used": {
                    "type": "string"
                }
            }
        },
        "service.HourlyStat": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "hour": {
                    "type": "integer"
                }
            }
        },
        "service.MemoryInfo": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "string"
                },
                "total": {
                    "type": "string"
                },
                "used": {
                    "type": "string"
                }
            }
        },
        "service.MethodStat": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                }
            }
        },
        "service.NetInfo": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "string"
                },
                "total": {
                    "type": "string"
                },
                "used": {
                    "type": "string"
                }
            }
        },
        "service.OperatorLogStatistics": {
            "type": "object",
            "properties": {
                "costDist": {
                    "$ref": "#/definitions/service.CostDist"
                },
                "hourlyStats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.HourlyStat"
                    }
                },
                "methodStats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.MethodStat"
                    }
                },
                "statusCodeStats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.StatusCodeStat"
                    }
                },
                "summary": {
                    "$ref": "#/definitions/service.DashboardSummary"
                }
            }
        },
        "service.StatusCodeStat": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "service.SystemResource": {
            "type": "object",
            "properties": {
                "cpu": {
                    "$ref": "#/definitions/service.CPUInfo"
                },
                "disk": {
                    "$ref": "#/definitions/service.DiskInfo"
                },
                "memory": {
                    "$ref": "#/definitions/service.MemoryInfo"
                },
                "net": {
                    "$ref": "#/definitions/service.NetInfo"
                }
            }
        },
...
    }
}